refactor: update test to DateInput
